requirements.txt
nonebot2
nonebot-adapter-kook
psmb-client
hitmc-messages
pydantic~=1.9.2

loguru~=0.6.0